test: split dashboard memory curation API coverage
